requirements.txt
py-clob-client==0.19.0
python-dotenv==1.0.0
web3==6.11.1
